+++ src/Dirent.cpp
@@ -3,8 +3,8 @@
 
 void LinuxDirentRemove(linux_dirent* dentsp, linux_dirent* d, int* cnt_dents_bytes)
 {
-    unsigned int dirent_end = (unsigned int)dentsp + *cnt_dents_bytes;
-    unsigned int next_dirnet = (unsigned int)d + d->d_reclen;
+    uintptr_t dirent_end = (uintptr_t)dentsp + *cnt_dents_bytes;
+    uintptr_t next_dirnet = (uintptr_t)d + d->d_reclen;
     
     *cnt_dents_bytes -= d->d_reclen;
     
@@ -13,8 +13,8 @@ void LinuxDirentRemove(linux_dirent* dentsp, linux_dirent* d, int* cnt_dents_byt
 
 void LinuxDirent64Remove(linux_dirent64* dentsp, linux_dirent64* d, int* cnt_dents_bytes)
 {
-    unsigned int dirent_end = (unsigned int)dentsp + *cnt_dents_bytes;
-    unsigned int next_dirnet = (unsigned int)d + d->d_reclen;
+    uintptr_t dirent_end = (uintptr_t)dentsp + *cnt_dents_bytes;
+    uintptr_t next_dirnet = (uintptr_t)d + d->d_reclen;
     
     *cnt_dents_bytes -= d->d_reclen;
